Optimization code

+++ ithub/users/serializers.py
@@ -2,11 +2,54 @@ from os import remove
 from shutil import rmtree
 
 from django.contrib.auth import authenticate
+from drf_writable_nested import WritableNestedModelSerializer
 from ithub.settings import MEDIA_ROOT
 from rest_framework import serializers
 from rest_framework.exceptions import ValidationError
 
-from .models import User
+from .models import User, Profile
+
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        exclude = ["last_login", "password"]
+        read_only_fields = (
+            'id', "data_joined", "is_active", "is_superuser",
+            "groups", "user_permissions"
+        )
+
+
+class AnotherUserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        exclude = ["last_login", "password", "email"]
+
+
+class MeProfileSerializer(WritableNestedModelSerializer, serializers.ModelSerializer):
+    user = UserSerializer(read_only=True)
+
+    def update(self, instance, validated_data):
+        if "avatar" in validated_data.keys():
+            if instance.avatar.url != "/media/default/avatar.png":
+                try:
+                    remove(instance.avatar.path)
+                except Exception as ex:
+                    print(ex)
+
+        return super(MeProfileSerializer, self).update(instance, validated_data)
+
+    class Meta:
+        model = Profile
+        fields = '__all__'
+
+
+class AnotherProfileSerializer(serializers.ModelSerializer):
+    user = AnotherUserSerializer()
+
+    class Meta:
+        model = Profile
+        fields = '__all__'
 
 
 class LoginSerializer(serializers.Serializer):
@@ -19,8 +62,8 @@ class LoginSerializer(serializers.Serializer):
             if user.is_active:
                 return user
 
-            raise serializers.ValidationError('Your account is not active')
-        raise serializers.ValidationError('Incorrect Credentials Passed.')
+            raise ValidationError('Your account is not active')
+        raise ValidationError('Incorrect Credentials Passed.')
 
 
 class CreateUserSerializer(serializers.ModelSerializer):
@@ -34,37 +77,17 @@ class CreateUserSerializer(serializers.ModelSerializer):
             validated_data["email"],
             validated_data['password']
         )
+        # Profile.objects.create(user=user)
         return user
 
 
-class MeUserSerializer(serializers.ModelSerializer):
-    is_active = serializers.BooleanField(read_only=True)
-    is_superuser = serializers.BooleanField(read_only=True)
-
-    def update(self, instance, validated_data):
-        if "avatar" in validated_data.keys():
-            print(instance.avatar.url)
-            if instance.avatar.url != "/media/default/avatar.png":
-                try:
-                    remove(instance.avatar.path)
-                except Exception as ex:
-                    print(ex)
-
-        new_inst = super(MeUserSerializer, self).update(instance, validated_data)
-        return new_inst
-
-    class Meta:
-        model = User
-        exclude = ["groups", "user_permissions", "password"]
 
 
-class AnotherUserSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        exclude = ["email", "groups", "user_permissions", "last_login", "password"]
 
+class DeleteUserSerializer(serializers.Serializer):
+    email = serializers.EmailField(max_length=100)
+    password = serializers.CharField(max_length=150)
 
-class DeleteUserSerializer(serializers.ModelSerializer):
     def validate(self, attrs):
         email = attrs.get("email", None)
         password = attrs.get("password", None)
@@ -75,17 +98,19 @@ class DeleteUserSerializer(serializers.ModelSerializer):
         if not instance.check_password(password):
             raise ValidationError("password incorrectly")
 
-        rmtree(MEDIA_ROOT / "users" / str(instance.id))
+        if instance.user.avatar.url != "/media/default/avatar.png":
+            try:
+                rmtree(MEDIA_ROOT / "users" / str(instance.id))
+            except Exception as ex:
+                print(ex)
+
         instance.delete()
 
-        return {"success": "true"}
-
-    class Meta:
-        model = User
-        fields = ["email", "password"]
+        return {"success": True}
 
 
-class PasswordResetSerializer(serializers.ModelSerializer):
+class PasswordResetSerializer(serializers.Serializer):
+    password = serializers.CharField(max_length=200)
     new_password = serializers.CharField(max_length=200)
     new_password_repeat = serializers.CharField(max_length=200)
 
@@ -94,19 +119,16 @@ class PasswordResetSerializer(serializers.ModelSerializer):
         instance = self.instance
 
         if not instance.check_password(password):
-            raise ValidationError("password incorrectly")
+            raise ValidationError("The current password is incorrect")
 
         new_password = attrs.get("new_password")
         new_password_repeat = attrs.get("new_password_repeat")
 
         if new_password != new_password_repeat or instance.check_password(new_password):
-            raise ValidationError("New password incorrectly")
+            raise ValidationError("The new password does not match the recurrence, or the new password is the same as "
+                                  "the old one")
 
         instance.set_password(new_password)
         instance.save()
 
-        return {"success": "true"}
-
-    class Meta:
-        model = User
-        fields = ["password", "new_password", "new_password_repeat"]
+        return {"success": True}
